# swagger.yaml
openapi: 3.0.0
info:
  title: 15-Jobs-Api
  contact: {}
  version: '1.0'
servers:
- url: https://jobs-api-15.onrender.com/api/v1
  variables: {}
paths:
...
  /jobs:
    get:
      tags:
      - Jobs
      summary: Get all jobs
      operationId: Getalljobs
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    post:
      tags:
      - Jobs
      summary: Create job
      operationId: Createjob
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreatejobRequest'
              - example:
                  position: front-end
                  company: zoho
            example:
              position: front-end
              company: zoho
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
  schemas:
...
    CreatejobRequest:
      title: CreatejobRequest
      required:
      - position
      - company
      type: object
      properties:
        position:
          type: string
        company:
          type: string
      example:
        position: front-end
        company: zoho
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...